First request to ChatGPT
Here was my first request to ChatGPT to get it to generate a
plugin generator specifically for generating speechhandler plugins
in the Naomi style:

"I would love your help in creating a tool in python to help people
create new speechhandler plugins for Naomi. Here is an example
speechhandler plugin:
https://github.com/aaronchantrill/NaomiNetworkCheck

"One of the big issues with the Naomi project is making it easy for
people to register their plugins with the project. Right now, you
have to create a plugin.info file with your plugin that includes
the url of the repository that will be hosting it (on github,
gitlab, etc.) then when you "publish" your plugin, you have to
create a new pull request where you insert a new line into the CSV
file at
https://github.com/NaomiProject/naomi-plugins/blob/master/plugins.csv
which includes a lot of the information in the plugin.info file,
plus the actual commit signature of the current version. This
information is hard to synchronize. What I would like would be a
GUI for adding a bunch of intent keywords and templates, then the
name, description and license to generate a simple plugin that
just repeats back what it hears. The user then edits the python in
whatever ide they like, then when they are ready to publish, they
reopen the GUI, select the project, and hit a publish button.
This will automatically collect the information from the
plugin.info file, push the current version to github, then make a
pull request with the new line in the CSV file. Does that make
sense?"

This pull request is what ChatGPT first came up with. I wasn't
specific about what packages to use for the GUI part, both
because I assumed it would use something normal like tkinter
and because I was interested to see what it would choose if it
chose something else. It chose to use PySimpleGUI. When I
tried to pip install it, I got a EULA informing me that the
package requires registration and for non-hobbyist users it
also requires a paid subscription. This was a non-starter for
me since I can imagine that business users might want to
create plugins, and forcing them to license a Python package
due to an arbitrary choice by ChatGPT just didn't seem
reasonable. Requiring users to register is also problematic
for me from a privacy perspective. It looks like PySimpleGUI
is in the process of shutting down anyway. I never actually
tried to run this version.
shutting down completely anyway.