Fixes to Fourth Request
A couple of these issues showed up when I first ran the code, but most of them
showed up when I tried testing the plugin that the code generated. The actual
process of pushing the generated plugin, forking the naomi-plugins repository,
adding a line to plugins.csv, pushing that to the forked repository and then
creating a pull request all appears to have gone fine except for the fact that
the master branch of my forked repository had diverged. I've since learned the
importance of always committing my changes to a temporary branch so I can keep
my master/main branch synchronized with upstream.

main.py - Added a shebang and set the execute bit in permissions so it can be
run directly rather than as a python argument. Added logging and fixed the
name of an imported function that I had to change.

plugin_editor.py - It bothered me that the background colors of the form
controls where inconsistent for no good reason. I decided I wanted to use
white for controls that the user can type directly into, and grey for
controls that are read-only or have to be modified using the side-buttons.

plugin_generator.py - This was where I found most of the issues, and mostly
with the plugin that the system generated rather than with the program itself
running.

First, it imported the SpeechHandlerPlugin class directly, which caused an
issue where the plugin system was only expecting one class to be directly
defined. I had to change it to import plugin rather than the
SpeechHandlerPlugin class directly.

Second, I really didn't like the way it formatted the intents it was defining.
It was using pprint.pformat for formatting the intent structure, but this was
doing odd stuff like inserting single quotes into the wrong places, causing
the resulting plugin to generate syntax errors. I much prefer the formatting
of json.dumps, so I switched to using that. I was also able to simplify the
generation code, as the program was doing unnecessary processing.

Third, a simple handle script like this should always return True. It would
only return False if the handler determined that it was unable to handle the
request. This is generally used with expect() and confirm().

The keywords were being defined with self.gettext() around them, which is not
necessary or desireable here because we are defining different locales, and
there may be more ways a user might choose to say something in one language
than another. GetText is more for translating interface labels to different
languages. This is probably because of one of the examples that I showed to
ChatGPT.

Finally, the request text is gotten using the 'input' property of the intent,
not the 'text' property.

publisher.py - the "Plugin" section and "Name" and "License" values must be
capitalized in order to be used by Naomi. Also, the repository url belongs
in the "URL" value, not "repo_url". Finally, ChatGPT both imported a
read_plugin_info function from storage and then defined a read_plugin_info
function that called it, resulting in an infinite loop/stack overflow.

This was an interesting experiment. Most of the mistakes that ChatGPT
made were similar to things I would expect a junior programmer to have
trouble with also, but I don't know that I can explain why things
should have been written differently to ChatGPT or if that would be
worthwhile. It was motivating to see the application come together
so quickly, and I enjoy debugging so I think it was helpful.